+++ src/pages/api/hooks/auth.ts
@@ -1,3 +1,5 @@
+import { User } from "@prisma/client";
+
 import { sendWelcomeEmail, sendWelcomeEmailBulk } from "@/pages/api/email/index";
 import { Prisma } from "@/pages/api/prisma";
 
@@ -12,4 +14,25 @@ const welcomeUserEmailHook = async (params: Prisma.MiddlewareParams) => {
   }
 };
 
-export { welcomeUserEmailHook };
+const sanitizeUserEmail = (params: Prisma.MiddlewareParams) => {
+  switch (params.action) {
+    case "create":
+      params.args.data.email = params.args.data.email.toLowerCase();
+      break;
+    case "update":
+      if ("email" in params.args.data) {
+        params.args.data.email.set = params.args.data.email.set.toLowerCase();
+      }
+      break;
+    case "createMany":
+      params.args.data = params.args.data.map((user: User) => ({
+        ...user,
+        email: user.email.toLowerCase(),
+      }));
+      break;
+  }
+
+  return params;
+};
+
+export { sanitizeUserEmail, welcomeUserEmailHook };